chore: app 모듈 이름 Bootstrap으로 변경
