+++ Bootstrap-Module/src/main/kotlin/com/asap/bootstrap/AppApplication.kt
@@ -1,0 +1,13 @@
+package com.asap.bootstrap
+
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
+
+@SpringBootApplication
+class AppApplication {
+
+}
+
+fun main(args: Array<String>) {
+    SpringApplication.run(AppApplication::class.java, *args)
+}
