+++ Bootstrap-Module/src/main/kotlin/com/asap/bootstrap/auth/dto/SocialLoginRequest.kt
@@ -1,0 +1,9 @@
+package com.asap.bootstrap.auth.dto
+
+import io.swagger.v3.oas.annotations.media.Schema
+
+@Schema(description = "소셜 로그인 요청")
+data class SocialLoginRequest(
+    @Schema(description = "oauth access token")
+    val accessToken: String,
+)
