+++ Bootstrap-Module/src/main/kotlin/com/asap/bootstrap/user/dto/RegisterUserResponse.kt
@@ -1,0 +1,12 @@
+package com.asap.bootstrap.user.dto
+
+import io.swagger.v3.oas.annotations.media.Schema
+
+@Schema(description = "회원 가입 응답")
+data class RegisterUserResponse(
+    @Schema(description = "access token")
+    val accessToken: String,
+    @Schema(description = "refresh token")
+    val refreshToken: String
+) {
+}
