--- app/src/main/kotlin/com/asap/app/auth/dto/SocialLoginRequest.kt
@@ -1,9 +1,0 @@
-package com.asap.app.auth.dto
-
-import io.swagger.v3.oas.annotations.media.Schema
-
-@Schema(description = "소셜 로그인 요청")
-data class SocialLoginRequest(
-    @Schema(description = "oauth access token")
-    val accessToken: String,
-)
