--- app/src/test/java/com/asap/app/user/controller/UserControllerTest.kt
@@ -1,63 +1,0 @@
-package com.asap.app.user.controller
-
-import com.asap.app.user.dto.RegisterUserRequest
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
-import org.junit.jupiter.api.Test
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
-import org.springframework.http.MediaType
-import org.springframework.test.web.servlet.MockMvc
-import org.springframework.test.web.servlet.post
-import java.time.LocalDate
-
-@WebMvcTest(UserController::class)
-@AutoConfigureMockMvc
-class UserControllerTest {
-
-    @Autowired
-    private lateinit var mockMvc: MockMvc
-
-    private val objectMapper: ObjectMapper = ObjectMapper().registerModules(JavaTimeModule())
-
-    @Test
-    fun registerUserTest(){
-        // given
-        val request = RegisterUserRequest("register", true, true, true, LocalDate.now())
-        // when
-        val response = mockMvc.post("/api/v1/users") {
-            contentType = MediaType.APPLICATION_JSON
-            content = objectMapper.writeValueAsString(request)
-        }
-        // then
-        response.andExpect {
-            status { isOk() }
-            jsonPath("$.accessToken") {
-                exists()
-                isString()
-                isNotEmpty()
-            }
-            jsonPath("$.refreshToken") {
-                exists()
-                isString()
-                isNotEmpty()
-            }
-        }
-    }
-
-    @Test
-    fun registerUserNotExistsRegisterTokenTest(){
-        // given
-        val request = RegisterUserRequest("nonExistsToken", false, true, true, LocalDate.now())
-        // when
-        val response = mockMvc.post("/api/v1/users") {
-            contentType = MediaType.APPLICATION_JSON
-            content = objectMapper.writeValueAsString(request)
-        }
-        // then
-        response.andExpect {
-            status { isBadRequest() }
-        }
-    }
-}
